EKLAMU-1539:flicker sensor bringup
Reviewed-on: https://gerrit-odm.mot.com/c/home/repo/dev/platform/android/quark/kernel_device_modules-6.6/+/4820
Tested-by: gerrito

+++ drivers/misc/mediatek/sensor/1.0/alsps/alsps.c
@@ -642,8 +642,8 @@ static ssize_t rearalscali_store(struct device *dev,
 	int err = 0;
 	uint8_t *cali_buf = NULL;
 
-	pr_info("rearals set cali:%d count=%d\n", *(int*)buf, count);
-	cali_buf = vzalloc(count);
+	pr_info("rearals set cali:%d count=%d\n", *(int*)buf, (int)count);
+	cali_buf = vzalloc((int)count);
 	if (!cali_buf)
 		return -ENOMEM;
 	memcpy(cali_buf, buf, count);
